Error in DateTime resolution for rounding (wrong logic gate). Error in EventType generation (still on-going). Removed errant use cases.

+++ NEsper/NEsper.Common/common/internal/util/SimpleNumberCoercerFactory.CoercerBigInt.cs
@@ -44,7 +44,7 @@ namespace com.espertech.esper.common.@internal.util
                 return numToCoerce.AsBigInteger();
             }
 
-            public Type ReturnType => typeof(long);
+            public Type ReturnType => typeof(BigInteger);
 
             public CodegenExpression CoerceCodegen(
                 CodegenExpression value,
